Clean: Relocate STEX BadAllocIfNull
