+++ src/imageproc/BadAllocIfNull.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "STEX_BadAllocIfNull.h"
+#include "BadAllocIfNull.h"
 #include <QImage>
 #include <new>
