+++ src/imageproc/BadAllocIfNull.h
